@@ -7,42 +7,44 @@
     "start": "npm run build-themes && ng serve",
     "start-jit": "npm run build-themes && ng serve --aot=false",
     "lint": "tslint -p src/tsconfig.json",
-    "test": "ng test",
+    "test": "npm run build-themes && ng test",
     "pree2e": "webdriver-manager update",
     "e2e": "protractor",
     "fetch-local": "bash ./tools/fetch-assets-local.sh",
     "build-themes": "bash ./tools/build-themes.sh",
     "prod-build": "npm run build-themes && ng build --aot --prod",
-    "postinstall": "webdriver-manager update && bash ./tools/build-themes.sh && bash ./tools/fetch-assets.sh",
+    "postinstall": "webdriver-manager update && bash ./tools/build-themes.sh",
     "publish-prod": "bash ./tools/deploy.sh stable prod",
-    "publish-dev": "bash ./tools/deploy.sh"
+    "publish-dev": "bash ./tools/deploy.sh",
+    "publish-beta": "bash ./tools/deploy.sh stable beta"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.9",
-    "@angular/cdk": "^6.4.6",
-    "@angular/common": "^6.0.9",
-    "@angular/compiler": "^6.0.9",
-    "@angular/core": "^6.0.9",
-    "@angular/forms": "^6.0.9",
-    "@angular/http": "^6.0.9",
-    "@angular/material": "^6.4.6",
-    "@angular/material-moment-adapter": "^6.4.6",
-    "@angular/platform-browser": "^6.0.9",
-    "@angular/platform-browser-dynamic": "^6.0.9",
-    "@angular/router": "^6.0.9",
+    "@angular/animations": "^7.0.0",
+    "@angular/cdk": "^7.2.0",
+    "@angular/common": "^7.0.0",
+    "@angular/compiler": "^7.0.0",
+    "@angular/core": "^7.0.0",
+    "@angular/forms": "^7.0.0",
+    "@angular/http": "^7.0.0",
+    "@angular/material": "^7.2.0",
+    "@angular/material-moment-adapter": "^7.2.0",
+    "@angular/platform-browser": "^7.0.0",
+    "@angular/platform-browser-dynamic": "^7.0.0",
+    "@angular/router": "^7.0.0",
     "core-js": "^2.5.6",
+    "hammerjs": "^2.0.8",
     "moment": "^2.22.1",
     "rxjs": "^6.1.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.6.8",
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^6.0.9",
-    "@types/jasmine": "^2.5.41",
+    "@angular-devkit/build-angular": "^0.10.1",
+    "@angular/cli": "^7.0.1",
+    "@angular/compiler-cli": "^7.0.0",
+    "@types/jasmine": "^2.8.9",
     "@types/node": "^10.0.9",
-    "firebase-tools": "^3.15.3",
+    "firebase-tools": "^6.0.0",
     "gulp": "^3.9.1",
     "gulp-replace": "^1.0.0",
     "hammerjs": "^2.0.8",
@@ -53,13 +55,15 @@
     "karma-browserstack-launcher": "^1.3.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "^1.1.1",
+    "karma-firefox-launcher": "^1.1.0",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-sauce-launcher": "^1.2.0",
+    "natives": "^1.1.6",
     "node-sass": "^4.5.3",
     "protractor": "~5.3.0",
     "ts-node": "^6.0.3",
     "tslint": "^5.10.0",
-    "typescript": "~2.7.2"
+    "typescript": "^3.1.3"
   }
 }
