@@ -1,6 +1,6 @@
 // Karma configuration file, see link for more information
 // https://karma-runner.github.io/0.13/config/configuration-file.html
-const {customLaunchers, platformMap} = require('./browser-providers');
+const {customLaunchers} = require('./karma-custom-launchers');
 
 module.exports = function (config) {
   config.set({
@@ -9,6 +9,7 @@ module.exports = function (config) {
     plugins: [
       require('karma-jasmine'),
       require('karma-chrome-launcher'),
+      require('karma-firefox-launcher'),
       require('karma-jasmine-html-reporter'),
       require('karma-coverage-istanbul-reporter'),
       require('@angular-devkit/build-angular/plugins/karma'),
@@ -28,51 +29,12 @@ module.exports = function (config) {
     colors: true,
     logLevel: config.LOG_INFO,
     autoWatch: true,
-    browsers: ['Chrome'],
+    browsers: ['ChromeHeadlessLocal'],
     singleRun: false,
     customLaunchers: customLaunchers,
-
-    sauceLabs: {
-      testName: 'material.angular.io',
-      startConnect: false,
-      recordVideo: false,
-      recordScreenshots: false,
-      options: {
-        'selenium-version': '2.48.2',
-        'command-timeout': 600,
-        'idle-timeout': 600,
-        'max-duration': 5400
-      }
-    },
-
-    browserStack: {
-      project: 'material.angular.io',
-      startTunnel: false,
-      retryLimit: 1,
-      timeout: 600,
-      pollingTimeout: 20000
-    },
   });
 
   if (process.env['TRAVIS']) {
-
-    let buildId = `TRAVIS #${process.env.TRAVIS_BUILD_NUMBER} (${process.env.TRAVIS_BUILD_ID})`;
-    let platformType = process.env.MODE;
-
-    if (platformType === 'saucelabs') {
-
-      config.sauceLabs.build = buildId;
-      config.sauceLabs.tunnelIdentifier = process.env.TRAVIS_JOB_NUMBER;
-
-    } else if (platformType === 'browserstack') {
-
-      config.browserStack.build = buildId;
-      config.browserStack.tunnelIdentifier = process.env.TRAVIS_JOB_NUMBER;
-
-    } else {
-      throw new Error(`Platform "${platform}" unknown, but Travis specified. Exiting.`);
-    }
-
-    config.browsers = platformMap[platformType];
+    config.browsers = ['ChromeHeadlessCI', 'FirefoxHeadless']
   }
 };
