@@ -8,25 +8,25 @@ import {MatNativeDateModule} from '@angular/material';
 import {ExampleModule} from '@angular/material-examples';
 
 import {MaterialDocsApp} from './material-docs-app';
-import {HomepageModule} from './pages/homepage/homepage';
+import {HomepageModule} from './pages/homepage';
 import {MATERIAL_DOCS_ROUTES} from './routes';
-import {ComponentListModule} from './pages/component-list/component-list';
+import {ComponentListModule} from './pages/component-list';
 import {ComponentViewerModule} from './pages/component-viewer/component-viewer';
 import {ComponentCategoryListModule} from './pages/component-category-list/component-category-list';
 import {ComponentSidenavModule} from './pages/component-sidenav/component-sidenav';
 import {FooterModule} from './shared/footer/footer';
 import {ComponentPageTitle} from './pages/page-title/page-title';
 import {ComponentHeaderModule} from './pages/component-page-header/component-page-header';
-import {StyleManager} from './shared/style-manager/style-manager';
+import {StyleManager} from './shared/style-manager';
 import {SvgViewerModule} from './shared/svg-viewer/svg-viewer';
-import {ThemePickerModule} from './shared/theme-picker/theme-picker';
-import {StackblitzButtonModule} from './shared/stackblitz/stackblitz-button';
-import {NavBarModule} from './shared/navbar/navbar';
+import {ThemePickerModule} from './shared/theme-picker';
+import {StackblitzButtonModule} from './shared/stackblitz';
+import {NavBarModule} from './shared/navbar';
 import {ThemeStorage} from './shared/theme-picker/theme-storage/theme-storage';
 import {GuideItems} from './shared/guide-items/guide-items';
 import {DocumentationItems} from './shared/documentation-items/documentation-items';
-import {GuideListModule} from './pages/guide-list/guide-list';
-import {GuideViewerModule} from './pages/guide-viewer/guide-viewer';
+import {GuideListModule} from './pages/guide-list';
+import {GuideViewerModule} from './pages/guide-viewer';
 import {DocViewerModule} from './shared/doc-viewer/doc-viewer-module';
 import {
   CanActivateComponentSidenav
