@@ -74,7 +74,7 @@ describe('StackblitzWriter', () => {
     }
     flushMicrotasks();
 
-    expect(form.elements.length).toBe(14);
+    expect(form.elements.length).toBe(15);
 
     // Should have correct tags
     expect(form.elements[0].getAttribute('name')).toBe('tags[0]');
@@ -96,9 +96,10 @@ describe('StackblitzWriter', () => {
     expect(form.elements[8].getAttribute('name')).toBe('files[polyfills.ts]');
     expect(form.elements[9].getAttribute('name')).toBe('files[.angular-cli.json]');
     expect(form.elements[10].getAttribute('name')).toBe('files[main.ts]');
-    expect(form.elements[11].getAttribute('name')).toBe('files[app/test.ts]');
-    expect(form.elements[12].getAttribute('name')).toBe('files[app/test.html]');
-    expect(form.elements[13].getAttribute('name')).toBe('files[app/src/detail.ts]');
+    expect(form.elements[11].getAttribute('name')).toBe('files[material-module.ts]');
+    expect(form.elements[12].getAttribute('name')).toBe('files[app/test.ts]');
+    expect(form.elements[13].getAttribute('name')).toBe('files[app/test.html]');
+    expect(form.elements[14].getAttribute('name')).toBe('files[app/src/detail.ts]');
   }));
 });
 
@@ -115,6 +116,7 @@ const TEST_URLS = [
   '/assets/stackblitz/polyfills.ts',
   '/assets/stackblitz/.angular-cli.json',
   '/assets/stackblitz/main.ts',
+  '/assets/stackblitz/material-module.ts',
   'http://material.angular.io/test.ts',
   'http://material.angular.io/test.html',
   'http://material.angular.io/src/detail.ts',
