@@ -10,17 +10,13 @@ describe('ThemePicker', () => {
     }).compileComponents();
   }));
 
-  it('should install theme based on href', () => {
+  it('should install theme based on name', () => {
     const fixture = TestBed.createComponent(ThemePicker);
     const component = fixture.componentInstance;
-    const href = 'pink-bluegrey.css';
+    const name = 'pink-bluegrey';
     spyOn(component.styleManager, 'setStyle');
-    component.installTheme({
-      primary: '#E91E63',
-      accent: '#607D8B',
-      href,
-    });
+    component.installTheme(name);
     expect(component.styleManager.setStyle).toHaveBeenCalled();
-    expect(component.styleManager.setStyle).toHaveBeenCalledWith('theme', `assets/${href}`);
+    expect(component.styleManager.setStyle).toHaveBeenCalledWith('theme', `assets/${name}.css`);
   });
 });
