@@ -1,11 +1,24 @@
-import {Component, ViewEncapsulation, ChangeDetectionStrategy, NgModule} from '@angular/core';
+import {
+  Component,
+  ViewEncapsulation,
+  ChangeDetectionStrategy,
+  NgModule,
+  OnInit,
+  OnDestroy,
+} from '@angular/core';
 import {StyleManager} from '../style-manager/style-manager';
 import {ThemeStorage, DocsSiteTheme} from './theme-storage/theme-storage';
 import {
-  MatButtonModule, MatGridListModule, MatIconModule, MatMenuModule,
-  MatTooltipModule
+  MatButtonModule,
+  MatGridListModule,
+  MatIconModule,
+  MatMenuModule,
+  MatTooltipModule,
 } from '@angular/material';
 import {CommonModule} from '@angular/common';
+import {ActivatedRoute} from '@angular/router';
+import {Subscription} from 'rxjs';
+import {map, filter} from 'rxjs/operators';
 
 
 @Component({
@@ -16,63 +29,73 @@ import {CommonModule} from '@angular/common';
   encapsulation: ViewEncapsulation.None,
   host: {'aria-hidden': 'true'},
 })
-export class ThemePicker {
-  currentTheme;
+export class ThemePicker implements OnInit, OnDestroy {
+  private _queryParamSubscription = Subscription.EMPTY;
+  currentTheme: DocsSiteTheme;
 
-  themes = [
+  themes: DocsSiteTheme[] = [
     {
       primary: '#673AB7',
       accent: '#FFC107',
-      href: 'deeppurple-amber.css',
+      name: 'deeppurple-amber',
       isDark: false,
     },
     {
       primary: '#3F51B5',
       accent: '#E91E63',
-      href: 'indigo-pink.css',
+      name: 'indigo-pink',
       isDark: false,
       isDefault: true,
     },
     {
       primary: '#E91E63',
       accent: '#607D8B',
-      href: 'pink-bluegrey.css',
+      name: 'pink-bluegrey',
       isDark: true,
     },
     {
       primary: '#9C27B0',
       accent: '#4CAF50',
-      href: 'purple-green.css',
+      name: 'purple-green',
       isDark: true,
     },
   ];
 
   constructor(
     public styleManager: StyleManager,
-    private _themeStorage: ThemeStorage
-  ) {
-    const currentTheme = this._themeStorage.getStoredTheme();
-    if (currentTheme) {
-      this.installTheme(currentTheme);
-    }
+    private _themeStorage: ThemeStorage,
+    private _activatedRoute: ActivatedRoute) {
+    this.installTheme(this._themeStorage.getStoredThemeName());
   }
 
-  installTheme(theme: DocsSiteTheme) {
-    this.currentTheme = this._getCurrentThemeFromHref(theme.href);
+  ngOnInit() {
+    this._queryParamSubscription = this._activatedRoute.queryParamMap
+      .pipe(map(params => params.get('theme')), filter(Boolean))
+      .subscribe(themeName => this.installTheme(themeName));
+  }
+
+  ngOnDestroy() {
+    this._queryParamSubscription.unsubscribe();
+  }
+
+  installTheme(themeName: string) {
+    const theme = this.themes.find(currentTheme => currentTheme.name === themeName);
+
+    if (!theme) {
+      return;
+    }
+
+    this.currentTheme = theme;
 
     if (theme.isDefault) {
       this.styleManager.removeStyle('theme');
     } else {
-      this.styleManager.setStyle('theme', `assets/${theme.href}`);
+      this.styleManager.setStyle('theme', `assets/${theme.name}.css`);
     }
 
     if (this.currentTheme) {
       this._themeStorage.storeTheme(this.currentTheme);
     }
-  }
-
-  private _getCurrentThemeFromHref(href: string): DocsSiteTheme {
-    return this.themes.find(theme => theme.href === href);
   }
 }
 
