@@ -1,7 +1,10 @@
 import {Component, NgModule} from '@angular/core';
+import {HttpClient} from '@angular/common/http';
 import {CommonModule} from '@angular/common';
-import {MatButtonModule, MatMenuModule} from '@angular/material';
-import {docVersions, materialVersion, VersionInfo} from '../version/version';
+import {MatButtonModule, MatIconModule, MatMenuModule} from '@angular/material';
+import {materialVersion, VersionInfo} from '../version/version';
+
+const versionUrl = 'https://material.angular.io/assets/versions.json';
 
 @Component({
   selector: 'version-picker',
@@ -11,7 +14,9 @@ export class VersionPicker {
   /** The currently running versin of Material. */
   materialVersion = materialVersion;
   /** The possible versions of the doc site. */
-  docVersions = docVersions;
+  docVersions = this.http.get(versionUrl);
+
+  constructor(private http: HttpClient) {}
 
   /** Updates the window location if the selected version is a different version. */
   onVersionChanged(version: VersionInfo) {
@@ -22,7 +27,7 @@ export class VersionPicker {
 }
 
 @NgModule({
-  imports: [MatButtonModule, MatMenuModule, CommonModule],
+  imports: [MatButtonModule, MatIconModule, MatMenuModule, CommonModule],
   exports: [VersionPicker],
   declarations: [VersionPicker]
 })
